requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.4
django-datetime-widget2==0.9.5
django-filter==2.4.0
django-simple-history==3.0.0
django-taggit==1.4.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
Markdown==3.3.4
parso==0.8.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
